Cbits cleanup extreme edition
Now uses a proper write buffer on the haskell end so that we can later
control how it is filled. Actual guaruntees about how the buffer is
copied into the C side of things are now possible. `BiFloat`s are now a
thing, to be replaced with something more flexible sometime,
unfortunately. No longer using the silly callback redirect scheme,
although that was really cute and hopefully there will be a use for it
sometime.